+++ js/views/messagesitem.js
@@ -20,6 +20,8 @@ define(function(require) {
 	var Marionette = require('backbone.marionette');
 	var Radio = require('radio');
 	var MessageTemplate = require('templates/message-list-item.html');
+	var imageplaceholder = require('views/imageplaceholder');
+	var tooltip = require('views/tooltip');
 
 	return Marionette.View.extend({
 		template: MessageTemplate,
@@ -35,10 +37,35 @@ define(function(require) {
 		modelEvents: {
 			change: 'render'
 		},
+
+		/**
+		 * Get the sender/recipient label as string
+		 *
+		 * @returns {String}
+		 */
+		_getMessageLabel: function() {
+			var sendRec = [];
+			if (this.model.folder.get('specialRole') === 'sent') {
+				sendRec = this.model.get('to');
+			} else {
+				sendRec = this.model.get('from');
+			}
+
+			switch (sendRec.length) {
+				case 0:
+					return '-';
+				case 1:
+					return sendRec[0].label;
+				default:
+					return sendRec[0].label + ' ' + t('mail', '& others');
+			}
+		},
+
 		serializeModel: function() {
 			var json = this.model.toJSON();
 			json.isUnified = require('state').currentAccount && require('state').currentAccount.get('isUnified');
 			json.sender = this.model.get('from')[0];
+			json.label = this._getMessageLabel();
 			return json;
 		},
 		onRender: function() {
@@ -55,7 +82,7 @@ define(function(require) {
 			if (displayName) {
 				_.each(this.$el.find('.avatar'), function(a) {
 					$(a).height('32px');
-					$(a).imageplaceholder(displayName, displayName);
+					imageplaceholder(a, displayName, displayName);
 				});
 			}
 
@@ -77,7 +104,7 @@ define(function(require) {
 				revert: 'invalid'
 			});
 
-			$('.action.delete').tooltip({placement: 'left'});
+			tooltip('.action.delete', {placement: 'left'});
 
 			this._fetchAvatar();
 		},