+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 0.7.9 – 2018-01-23
+### Fixed
+- Undefined variable warning in nextcloud.log
+- Inconsistent spelling
+- Saving attachments to Nextcloud files
+
+## 0.7.8 – 2018-01-15
+### Fixed
+- Loading of text messages
+- HTML to text conversion for replies
+- Recipient name rendering in sent folder
+
+## 0.7.7 – 2018-01-09
+### Added
+- New and updated translations
+- Editable account settings
+
 ## 0.7.6 – 2017-12-12
 ### Added
 - Opt-out for external avatars (Gravatar, favicon)